Removed filter directory and its classes

+++ src/main/java/com/atomtechnologies/ecommerce/demo/domain/Category.java
@@ -16,7 +16,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private int id;
 
     @Column(name = "name", nullable = false, unique = true)
     private String name;
@@ -25,4 +25,5 @@ public class Category {
     private String description;
 
     @OneToMany(mappedBy = "category", cascade = CascadeType.ALL)
-    private List<Product> products; }
+    private List<Product> products;
+}
