+++ src/main/java/com/atomtechnologies/ecommerce/demo/fetcher/CategoryFetcher.java
@@ -1,8 +1,8 @@
 package com.atomtechnologies.ecommerce.demo.fetcher;
 
 import com.atomtechnologies.ecommerce.demo.domain.Category;
-import com.atomtechnologies.ecommerce.demo.filter.GetCategoriesInput;
 import com.atomtechnologies.ecommerce.demo.service.CategoryService;
+import com.atomtechnologies.ecommerce.demo.types.GetCategoriesInput;
 import com.netflix.graphql.dgs.DgsComponent;
 import com.netflix.graphql.dgs.DgsQuery;
 import com.netflix.graphql.dgs.InputArgument;
@@ -21,8 +21,9 @@ public class CategoryFetcher {
     public List<Category> getCategories(@InputArgument GetCategoriesInput input) {
         return categoryService.getCategories(input.getPage(), input.getSize());
     }
+
     @DgsQuery
-    public Category getCategoryById(@InputArgument Long categoryId) {
+    public Category getCategoryById(@InputArgument int categoryId) {
         return categoryService.getCategoryById(categoryId);
     }
 }
