+++ src/main/java/com/atomtechnologies/ecommerce/demo/persistence/ReviewRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 @Repository
-public interface ReviewRepository extends JpaRepository<Review, Long> {
-    Page<Review> findByProductId(Long productId, Pageable pageable);
+public interface ReviewRepository extends JpaRepository<Review, Integer> {
+    Page<Review> findByProductId(int productId, Pageable pageable);
 }
